feat: update tasks handlers
